multi-agent-chat-app/backend/requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==0.21.0
openai==0.27.0
flask-cors==3.0.10